+++ assets/javascripts/kitten/components/lines/info-lines.js
@@ -1,7 +1,7 @@
-import React from 'react'
+import React, { Component } from 'react'
 import classNames from 'classnames'
 
-export class InfoLines extends React.Component {
+export class InfoLines extends Component {
   constructor(props) {
     super(props)
 
@@ -13,10 +13,15 @@ export class InfoLines extends React.Component {
   }
 
   renderInfo(element) {
-    const { key, value, id } = element
+    const { key, value, id, className, ...others } = element
+
+    const infoClassName = classNames(
+      'k-InfoLines__line',
+      className,
+    )
 
     return (
-      <div className="k-InfoLines__line" key={ id }>
+      <div key={ id } className={ infoClassName } { ...others }>
         <div className="k-InfoLines__line__key">
           { key }
         </div>
@@ -29,8 +34,25 @@ export class InfoLines extends React.Component {
   }
 
   render() {
+    const {
+      infos,
+      className,
+      withoutTopBottomBorder,
+      borderColorPrimary4,
+      ...others,
+    } = this.props
+
+    const infoLinesClassName = classNames(
+      'k-InfoLines',
+      {
+        'k-InfoLines--withoutTopBottomBorder': withoutTopBottomBorder,
+        'k-InfoLines--borderColorPrimary4': borderColorPrimary4,
+      },
+      className,
+    )
+
     return (
-      <div className="k-InfoLines">
+      <div className={ infoLinesClassName } { ...others }>
         { this.renderInfos() }
       </div>
     )
@@ -38,5 +60,8 @@ export class InfoLines extends React.Component {
 }
 
 InfoLines.defaultProps = {
+  className: null,
   infos: [], // Eg: [{ key: …, value: …, id: … }]
+  withoutTopBottomBorder: false,
+  borderColorPrimary4: false,
 }